Fix nested indexing/map access: skip borrow on bare expression results
Add isExpressionResultBare() helper to detect when an expression produces
a value not wrapped in Rc<RefCell>/Arc<Mutex> (e.g., index results, range
vars, bare vars). Use it in IndexExpr, map access, and len() to avoid
emitting incorrect .borrow()/.lock() calls on bare types.

This fixes nested access patterns like matrix[1][1] and inventory["k1"]["k2"]
which previously generated .borrow() on Vec/BTreeMap types.

Also includes re-transpiled test outputs and errors_custom promotion from
previous session.

+++ tests/XFAIL/generics_basic/main.rs
@@ -56,7 +56,7 @@ impl Unknown {
 
 pub fn map_keys(m: Rc<RefCell<Option<BTreeMap<K, Rc<RefCell<Option<V>>>>>>>) -> Rc<RefCell<Option<Vec<K>>>> {
 
-    let mut r = Rc::new(RefCell::new(Some(Vec::with_capacity(((*m.borrow().as_ref().unwrap()).len() as i32)))));
+    let mut r = Rc::new(RefCell::new(Some(Vec::with_capacity((*m.borrow().as_ref().unwrap()).len()))));
     for (k, _) in (*m.borrow().as_ref().unwrap()).clone() {
         {(*r.borrow_mut().as_mut().unwrap()).push(k); r.clone()};
     }
